Updated XML functions
removeLink/node and parseLinks/Nodes functions created.
the parse functions do not return a tuple as they are supposed to.
Instead, they return all the items in the link or node, in order.  So
IOT handle this, each item needs to be type checked, or the function
needs to be changed